+++ example/bounce/view/board.dart
@@ -1,18 +1,20 @@
 part of bounce;
 
 class Board extends Surface {
-  var movablePieces = new MovablePieces(32);
+  var movablePieces;
 
   Board(CanvasElement canvas): super(canvas) {
+    movablePieces = new MovablePieces(32,
+        distance: new Distance.from(canvas.width, canvas.height));
     movablePieces.randomInit();
-    movablePieces.forEach((MovablePiece mp) => mp.shape = PieceShape.CIRCLE);
+    movablePieces.forEach((MovablePiece p) => p.shape = PieceShape.CIRCLE);
   }
 
   draw() {
     clear();
-    movablePieces.forEach((MovablePiece mp) {
-      mp.move();
-      drawPiece(mp);
+    movablePieces.forEach((MovablePiece p) {
+      p.move();
+      drawPiece(p);
     });
   }
 }