+++ example/memory/view/board.dart
@@ -29,7 +29,8 @@ class Board extends Surface {
   draw() {
     super.draw();
     if (memory.recalled) { // game over
-      new Tag(canvas, cellSize * 2, cellSize * 2, 32, 'YOU WIN', color: 'red').draw();
+      new Tag(canvas, cellSize * 2, cellSize * 2, 32, 'YOU WIN',
+          color: 'red').draw();
     }
   }
 }